changing default values, reparing filters
git-svn-id: https://sccn.ucsd.edu/svn/software/eeglab@9679 1e58238c-3a72-4c0f-aabe-36d030380f39

+++ plugins/ERPLABfilters1.00/ERPLAB_butter1.m
@@ -89,7 +89,7 @@ smoothdata = zeros(numchan,frames);
 hflt = waitbar(0,'Filtering...');
 disp('Filtering input data...')
 for i = 1:numchan
-    smoothdata(i,:) = filtfilt(b,a, data(i,:)); % change to filtfilt for non-causal
+    smoothdata(i,:) = filtfilt(b,a, double(data(i,:))); % change to filtfilt for non-causal
     waitbar(i/numchan)
 end
 close(hflt)
